Remove anemic assets class.
