+++ src/Events/Editor/Full_Site/Provider.php
@@ -3,6 +3,7 @@
 namespace TEC\Events\Editor\Full_Site;
 
 use Tribe\Events\Editor\Blocks\Archive_Events;
+use \Tribe__Events__Main as Events_Main;
 
 /**
  * Class Provider
@@ -31,7 +32,7 @@ class Provider extends \tad_DI52_ServiceProvider {
 		$this->container->register(Hooks::class);
 
 		// Register the Service Provider for Assets.
-		$this->container->register( Assets::class );
+		$this->register_assets();
 
 		// Register the service provider itself on the container.
 		$this->container->singleton( static::class, $this );
@@ -39,5 +40,49 @@ class Provider extends \tad_DI52_ServiceProvider {
 
 	protected function register_singletons() {
 		$this->container->singleton( Archive_Events::class, Archive_Events::class, [ 'load' ] );
+	}
+
+	/**
+	 * Binds and sets up implementations.
+	 *
+     * @since TBD
+	 */
+	public function register_assets() {
+		$plugin = Events_Main::instance();
+
+		tribe_asset(
+			$plugin,
+			'tec-events-full-site',
+			'app/full-site.js',
+			[
+				'react',
+				'react-dom',
+				'wp-components',
+				'wp-api',
+				'wp-api-request',
+				'wp-blocks',
+				'wp-i18n',
+				'wp-element',
+				'wp-editor',
+				'tribe-common-gutenberg-data',
+				'tribe-common-gutenberg-utils',
+				'tribe-common-gutenberg-store',
+				'tribe-common-gutenberg-icons',
+				'tribe-common-gutenberg-hoc',
+				'tribe-common-gutenberg-elements',
+				'tribe-common-gutenberg-components',
+			],
+			'enqueue_block_editor_assets',
+			[
+				'in_footer'    => false,
+				'localize'     => [],
+				'conditionals' => [ $this, 'is_full_site_editor' ],
+				'priority'     => 106,
+			]
+		);
+	}
+
+	public function is_full_site_editor() {
+		return tec_is_full_site_editor();
 	}
 }
